# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FenixCRM API
  version: 1.0.0
paths:
...
  /api/v1/workflows/preview:
    post:
      summary: Preview workflow tooling diagnostics and visual graph from draft sources
      x-fr-traces:
      - FR-241
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowPreviewRequest'
      responses:
        '200':
          description: Preview passed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPreviewEnvelope'
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Preview has validation diagnostics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPreviewEnvelope'
        default:
          description: Unexpected response
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkflowPreviewRequest:
      type: object
      required:
      - dsl_source
      properties:
        dsl_source:
          type: string
        spec_source:
          type: string
...
    WorkflowPreviewEnvelope:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/WorkflowPreviewResponse'
    WorkflowPreviewResponse:
      type: object
      required:
      - passed
      - diagnostics
      - conformance
      - visual_graph
      properties:
        passed:
          type: boolean
        diagnostics:
          $ref: '#/components/schemas/WorkflowValidationDiagnostics'
        conformance:
          $ref: '#/components/schemas/WorkflowConformance'
        visual_graph:
          $ref: '#/components/schemas/WorkflowVisualProjection'
    WorkflowValidationDiagnostics:
      type: object
      properties:
        violations:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowDiagnostic'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowDiagnostic'
    WorkflowDiagnostic:
      type: object
      properties:
        checkId:
          type: integer
        code:
          type: string
        type:
          type: string
        description:
          type: string
        location:
          type: string
        line:
          type: integer
        column:
          type: integer
    WorkflowConformance:
      type: object
      required:
      - profile
      properties:
        profile:
          type: string
          enum: [safe, extended, invalid]
        details:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowConformanceDetail'
    WorkflowConformanceDetail:
      type: object
      required:
      - code
      - severity
      - message
      properties:
        code:
          type: string
        severity:
          type: string
          enum: [info, warning, error]
        message:
          type: string
        line:
          type: integer
        column:
          type: integer
...
    WorkflowVisualProjection:
      type: object
      required:
      - nodes
      - edges
      - conformance
      properties:
        workflow_name:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowVisualNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowVisualEdge'
        conformance:
          $ref: '#/components/schemas/WorkflowConformance'
    WorkflowVisualNode:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        label:
          type: string
        color:
          type: string
        position:
          $ref: '#/components/schemas/WorkflowVisualPosition'
        source:
          type: string
        effect:
          type: string
        properties:
          type: object
          additionalProperties: true
    WorkflowVisualPosition:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
    WorkflowVisualEdge:
      type: object
      properties:
        id:
          type: string
        from:
          type: string
        to:
          type: string
        connection_type:
          type: string
        properties:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
...